Fix compilation warning on some kernel versions
`atomic64_read` returns `long int` but not `long long int` on the
kernel 4.9 and older for amd64 and even on some newer kernels for arm64.

See https://lore.kernel.org/lkml/CAK8P3a1PpgjRKFemEi7mLuECh1srs8bbzoc07RSrFsx+febjaQ@mail.gmail.com/T/

+++ src/elastio-snap.c
@@ -3372,7 +3372,7 @@ static inline void wait_for_bio_complete(struct snap_device *dev)
 			atomic64_read(&dev->sd_submitted_cnt) == atomic64_read(&dev->sd_processed_cnt),
 			msecs_to_jiffies(WAIT_SUBMITTED_BIOS_MSEC))) {
 		LOG_WARN("failed wait for all submitted BIOs to be processed after %d ms. bio submitted = %lld, bio processed = %lld",
-				WAIT_SUBMITTED_BIOS_MSEC, atomic64_read(&dev->sd_submitted_cnt), atomic64_read(&dev->sd_processed_cnt));
+				WAIT_SUBMITTED_BIOS_MSEC, (u64)atomic64_read(&dev->sd_submitted_cnt), (u64)atomic64_read(&dev->sd_processed_cnt));
 	}
 }
